% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{lc}
\alias{lc}
\title{List comprehension}
\usage{
lc(expr, ...)
}
\arguments{
\item{expr}{An expression that will be evaluated for each element in the input lists.}

\item{...}{Additional list or predicate arguments.}
}
\description{
List comprehension
}
